refactor: change avatar path
