+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './Alert';
 export * from './Accordion';
-export * from './avatar/Avatar';
+export * from './Avatar';
 export * from './AvatarGroup';
 export * from './Badge';
 export * from './Breadcrumb';
